SK-1641 updated the sameple and  README
